EMSUSD-204 Material binding strength in AE
- Move the custom control class to its own file.
- Add an option menu to control the material binding strength in the attribute editor (AE).
- Change the binding strength when the user select a new value in the option menu.
- Hide the menu when the material in inherited and there is no direct material.
- Disable the menu when there is a stronger inherited material and a direct material.
- Show a tooltip that explain why it is disabled.
- Update children when material strength is changed.
- Don't change selection when going to the prim from which the material in inherited.
- Don't use italics.
