+++ lib/mayaUsd/resources/ae/CMakeLists.txt
@@ -19,8 +19,13 @@ install(FILES __init__.py DESTINATION ${CMAKE_INSTALL_PREFIX}/lib/python/ufe_ae/
 set(MAYAUSD_AE_TEMPLATES usdschemabase)
 
 foreach(_SUBDIR ${MAYAUSD_AE_TEMPLATES})
-    install(FILES ${_SUBDIR}/__init__.py  ${_SUBDIR}/custom_image_control.py ${_SUBDIR}/custom_enum_control.py ${_SUBDIR}/ae_template.py
+    install(FILES
+        ${_SUBDIR}/__init__.py
+        ${_SUBDIR}/custom_image_control.py
+        ${_SUBDIR}/custom_enum_control.py
+        ${_SUBDIR}/ae_template.py
         ${_SUBDIR}/attribute_custom_control.py
+        ${_SUBDIR}/material_custom_control.py
         DESTINATION ${CMAKE_INSTALL_PREFIX}/lib/python/ufe_ae/usd/nodes/${_SUBDIR}
     )
 endforeach()
